ml-model/requirements.txt
# ml-model/requirements.txt
# Install with:  pip install -r ml-model/requirements.txt

torch>=2.0.0
transformers>=4.38.0
Pillow>=10.0.0
sentencepiece>=0.1.99     # required by DonutProcessor tokeniser
